third_party/blink/web_tests/wpt_internal/css/css-anchor-position/inset-area-anchor-outside.html

<!DOCTYPE html>
<title>CSS Anchor Positioning: inset-area positioning - anchor outside containing block</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position/#inset-area">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/9598#issuecomment-1836854109">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #container {
    position: relative;
    width: 400px;
    height: 400px;
  }
  #anchor {
    position: absolute;
    left: -200px;
    top: 500px;
    width: 100px;
    height: 100px;
    anchor-name: --anchor;
  }
  #anchored {
    position: absolute;
    align-self: stretch;
    justify-self: stretch;
    position-anchor: --anchor;
  }
</style>
<div id="container">
  <div id="anchor"></div>
  <div id="anchored"></div>
</div>
<script>
  function test_inset_area(inset_area, expected_offsets) {
    anchored.style.insetArea = inset_area;

    test(() => {
      assert_equals(anchored.offsetLeft, expected_offsets.left, "Check expected offsetLeft");
      assert_equals(anchored.offsetTop, expected_offsets.top, "Check expected offsetTop");
      assert_equals(anchored.offsetWidth, expected_offsets.width, "Check expected offsetWidth");
      assert_equals(anchored.offsetHeight, expected_offsets.height, "Check expected offsetHeight");
    }, "Offsets for inset-area: " + inset_area);
  }

  test_inset_area("all", {left:0, top:0, width:400, height:400});

  test_inset_area("left / all", {left:-200, top:0, width:0, height:400});
  test_inset_area("left center / all", {left:-100, top:0, width:0, height:400});
  test_inset_area("center / top bottom", {left:-200, top:0, width:100, height:400});
  test_inset_area("center right / all", {left:-200, top:0, width:600, height:400});
  test_inset_area("right / all", {left:-100, top:0, width:500, height:400});

  test_inset_area("top / all", {left:0, top:0, width:400, height:500});
  test_inset_area("top center / all", {left:0, top:0, width:400, height:600});
  test_inset_area("center / left right", {left:0, top:500, width:400, height:100});
  test_inset_area("center bottom / all", {left:0, top:500, width:400, height:0});
  test_inset_area("bottom / all", {left:0, top:600, width:400, height:0});
</script>
